radio-transcription/requirements.txt
# Core dependencies
openai>=1.0.0                # GPT-4o API
python-dotenv>=1.0.0         # Environment variables
pydantic>=2.0.0             # Data validation

# Audio processing
sounddevice>=0.4.6          # Audio capture
soundfile>=0.12.0           # Audio file handling
numpy>=1.24.0               # Audio processing

# API server (for frontend integration)
fastapi>=0.104.0            # REST API
uvicorn>=0.24.0             # ASGI server
websockets>=12.0            # Real-time streaming
python-multipart>=0.0.6     # File uploads

# Logging and monitoring
loguru>=0.7.0               # Better logging
python-json-logger>=2.0.0   # JSON logs

# Optional but recommended
aiofiles>=23.0.0            # Async file operations
pydub>=0.25.1              # Audio format conversion (MP3, M4A, etc.)
requests>=2.31.0           # HTTP library (for launcher)
